+++ CHANGELOG.md
@@ -6,6 +6,17 @@
 ### Removed 
 ### Security
 
+# 2019-03-12-1
+### Added
+* Added task to import facilities for a single district 
+### Changed
+### Fixed
+### Deprecated 
+### Removed 
+* Facility creation via Card Reader
+### Security
+* Update project dependencies to fix security issues
+
 # 2019-02-18-1
 ### Added
 * Added a panel for showing a sync stats for Patient in each Facility & District